re-implements course category customization selectors.
previously, this was built using hinky custom selector components with
a metric ton of code, including running db queries in the form builder,
to make it all work.
note that that a list of checkboxes may not scale up to a huge number of
course categories, so this may have to switched again to different form
component, like a multi-select drop-down box (to be seen).