Sets block-drawer to be open by default.
Moodle 3 has a fixed, right-hand rail for blocks.
In its place, Moodle 4 has a collapsible drawer for blocks.
That block-drawer is/was closed by default.
This tripped up some of our users when evaluating Moodle 4 in preparation for the upgrade.
"My blocks are gone!!"
In order to make this transition less jarring, let's make the drawer
opened by default.