phpcs fix: the banner alerts callback script requires an authenticated user.
PHPCS complains that if you load the config file then an authenticated user
is also required (violation code: moodle.Files.RequireLogin.Missing).
This requires us to address the edge case of banner alerts appearing to
unauthenticated users on the frontpage. Problem solved by removing the
"dismiss alert" button from banners altogether if the user is not logged
in.