fix phpdoc linter errors.
Moodle is Very Opinionated (TM) as to what and how code should be
documented. Let's align with that.

While at it, I fixed up some class member references that `phpcbf` "forgot"
to update when it did its thing a few commits ago.

Also, I dropped the year from the copyright annotation everwhere - it's
always going to be the Regent's code, no matter what year it is.
This required the re-minification of some JS assets.